Updated cnvGetCounts & moved away from S4/Bioconductor objects
A lot of time was spent evaluating the incorporation/exenstion of
existing S4 sturctues from Bioconductor. The options did not have the flexibility
desired, so it was decided to stick fully with the data.table package and test for valid inputs.
See the new validObjects.R. The only remaining Bioconductor objects are for reading bam files
using the Rsamtools package.

Additionally, many helper functions were removed to focus on doing less things
better/more clearly.